Clean up toolbars, browser.
Replaced the red delete button / white "open page" button with toolbar buttons, which should be easier to discover since you don't have to scroll down.
Gave the in-app browser for detail pages a toolbar with back/refresh/stop/fwd,
and moved the action button to bottom.
Hid actionsheet in BrowserHelper as an implementation detail...