Omit optional 'isValid' field
The 'isValid' field is an optional flag that indicates whether the event passed
validation on the client. We re-validate on the server anyhow, and it's the
server-side validation counts. 'isValid' is merely a useful indicator of the
state of deployed client code. (I should probably rename it 'clientValidated'
or drop it altogether -- suggestions welcome.)

At any rate, it's an optional field, and since you're not validating things on
your end (which is fine!) you ought to simply omit it, I think.

This made my day, by the way!