Separate settings page. Smooth animations for login page keyboard reveal on iPad and iPhone
Move the debug mode setting to a new settings page which is accessed from the gear icon on the my uploads page. When keyboard is shown on the login page the logo and login container now swap places to ensure the keyboard is not covering any part of the login text boxes or login button. Works very smoothly on iPad and on small screens even when in landscape mode. Also involved some scaling of the logo on small screens when in landscape. Positioning is done in resolution independent manner.