Adjustment to the logo when keyboard is visible
Made the logo and login container no longer swap positions when the keyboard appeared. This left a lot of empty space on iPads. Instead, when the keyboard appears the following animation happens:
 - login container moves to logo position
 - logo alpha fades
 - logo programatically switches to black and white
 - logo size increases slightly
Animation is reversed when keyboard hides. The overall effect is very nice and results in less lonely whitespace on iPads.