Improved about page
Better compatibility with different screen sizes. Moved about button to left and added button to bottom left of My Uploads page. Made the Source section toggle to reveal source and license links for the iOS code base and also for other components used - this section should be easy to ammend for future additions.