Fixed alpha sort issue with browser list
Had been swapping the selected browser with the first browser in the list to cause the selection to appear on top, but this caused the items below the first one to no longer be alpha sorted. No more swapping - the selection is now removed and re-added to the front of the array thus leaving alpha sort of remaining browsers correct.