Observer bug causing crash when selecting image from gallery
Oddly, MyUploadsViewController's viewWillDisappear seems to be called without a matching viewDidAppear when an image is picked from the gallery. This seems to be an iOS thing (I think?) as the gallery selection code is baked-in. This causes a problem as the observer is added in the viewDidAppear and cannot be removed unless it was added or you'll get an exception.