Eliminated Getting Started page control dot update delay
One of the 3 page control dots at the bottom of the screen is lighter in color to indicate the current page. Previously the dot was being updated only after scrolling to new page had completed. Now it updates as soon as a majority of the screen is filled with a given page.