Picture of the day cycles between all cached pics of the day
Also normalized bundling of a default pic of the day. To bundle a different default image just remove the existing cache file from the bundle (the presently bundled cache file is named "POTD-2013-05-24" - note that the cache file contains more than just the image - it contains attribution data as well) and bundle a different one and change the BUNDLED_PIC_OF_DAY_DATE define to reflect the date of the newly bundled file. Cached filescan be grabbed from the app's "potd" folder - and just change the PIC_OF_THE_DAY_TO_DOWNLOAD_DAYS_AGO define to force it to download and create a cache file for a particular image if you want a new one to bundle.