Added license to picture of day attribution box.
Made the AspectFillThumbFetcher also cache image description and categories using the request it was already making. Then made it search these categories for commonly used known licenses. Caches which license it finds using "license" and "licenseurl" keys so license info can later be displayed or used. Presently using this license data in the image of the day attribution box.