Made pic of day attribution handle license not retrieved.
Says "Tap for License" for now if it couldn't extract the license name from the categories. Also, in the case that the license could not be determined the license url is also not known, so for now when this happens just made the license button link to the image's wiki page. (The wiki page should have license info). These are less than ideal stopgap solutions.